log4j要學 targetDaO差一點可以從SQL查但無法從JAVA接

+++ src/main/java/com/example/elandweb/dto/PageDto.java
@@ -13,5 +13,5 @@ import lombok.experimental.SuperBuilder;
 public class PageDto {
     private int page;
     private int size;
-    private int total;
+    private long total;
 }
